+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scores/Child/WorkoutSubmission/Feature/WorkoutSubmissionFeature+Action.swift
@@ -1,0 +1,26 @@
+//
+//  WorkoutSubmissionFeature+Action.swift
+//  StepTrackerTCA
+//
+//  Created by Sebastian Sciuba on 12/04/2025.
+//
+
+import ComposableArchitecture
+
+/// Implementation of `WorkoutSubmissionFeature` action
+extension WorkoutSubmissionFeature {
+    
+    @CasePathable
+    enum Action: ViewAction {
+        
+        case addValue
+        
+        case view(View)
+        
+        enum View {
+            case viewDidAppear
+            case add
+        }
+    }
+    
+}
