+++ StepTrackerTCA/Features/Records/ChildFeature/WorkoutsSummary/ChildFeature/Scores/Utilities/ScoresFeatureServicesKey.swift
@@ -17,7 +17,6 @@ private enum ScoresFeatureServicesKey: DependencyKey {
     })
 }
 
-
 extension DependencyValues {
     
     var scoresFeatureServices: ScoresFeatureServices {
